fix : useGatherings 파일 형식 변경
